Create Secure password checker
The pwned_api_check function implements a secure password checker that determines the number of times a given password has been compromised in known data breaches. The function employs a robust approach by securely hashing the password and then querying an API to retrieve relevant data. By utilizing the SHA-1 encryption algorithm and comparing a portion of the hashed password, it ensures the confidentiality of the original password while still providing accurate breach information. The function returns the count of password leaks to offer users insight into the security of their chosen password.